+++ src/components/NoteForm.js
@@ -1,11 +1,29 @@
-const NoteForm = ({ onSubmit, handleChange, value}) => {
+import { useState } from 'react' 
+
+const NoteForm = ({ createNote }) => {
+  const [newNote, setNewNote] = useState('') 
+
+  const handleChange = (event) => {
+    setNewNote(event.target.value)
+  }
+
+  const addNote = (event) => {
+    event.preventDefault()
+    createNote({
+      content: newNote,
+      important: Math.random() > 0.5,
+    })
+
+    setNewNote('')
+  }
+
   return (
     <div>
       <h2>Create a new note</h2>
 
-      <form onSubmit={onSubmit}>
+      <form onSubmit={addNote}>
         <input
-          value={value}
+          value={newNote}
           onChange={handleChange}
         />
         <button type="submit">save</button>
